pyproject.toml
[project]
name = "pnv3"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.11.18",
    "aiosqlite>=0.21.0",
    "argon2-cffi>=23.1.0",
    "fastapi>=0.115.12",
    "jinja2>=3.1.6",
    "uvicorn>=0.34.2",
]
